Allowing direct imports of models.
For example, `from intercom import User` instead of
`from intercom.user import User`.
