+++ intercom/api_operations/count.py
@@ -1,4 +1,3 @@
-from intercom import Intercom
 from intercom import utils
 
 
@@ -6,5 +5,6 @@ class Count(object):
 
     @classmethod
     def count(cls):
+        from intercom import Intercom
         response = Intercom.get("/counts/")
         return response[utils.resource_class_to_name(cls)]['count']
