+++ intercom/extended_api_operations/reply.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 
-from intercom import Intercom
 from intercom import utils
 
 
@@ -8,6 +7,7 @@ class Reply(object):
 
     @property
     def reply(self, reply_data):
+        from intercom import Intercom
         collection = utils.resource_class_to_collection_name(self.__class__)
         url = "/%s/%s/reply" % (collection, self.id)
         reply_data['conversation_id'] = self.id
