+++ tests/unit/company_spec.py
@@ -3,7 +3,7 @@ import json
 import re
 from describe import expect
 import intercom
-from intercom.company import Company
+from intercom import Company
 
 get = httpretty.GET
 r = re.compile
