+++ tests/unit/event_spec.py
@@ -4,8 +4,8 @@ import re
 import time
 from datetime import datetime
 from describe import expect
-from intercom.user import User
-from intercom.event import Event
+from intercom import User
+from intercom import Event
 
 post = httpretty.POST
 r = re.compile
