+++ tests/unit/tag_spec.py
@@ -2,7 +2,7 @@ import httpretty
 import json
 import re
 from describe import expect
-from intercom.tag import Tag
+from intercom import Tag
 from tests.unit import test_tag
 
 get = httpretty.GET
